refactor: Switched from double to float for performance

+++ Library/Vec3.cpp
@@ -28,37 +28,37 @@ Vec3& Vec3::operator/=(const Vec3 &v) {
     return *this;
 }
 
-Vec3& Vec3::operator+=(const double d) {
+Vec3& Vec3::operator+=(const float d) {
     vector[0] += d;
     vector[1] += d;
     vector[2] += d;
     return *this;
 }
 
-Vec3& Vec3::operator-=(const double d) {
+Vec3& Vec3::operator-=(const float d) {
     vector[0] -= d;
     vector[1] -= d;
     vector[2] -= d;
     return *this;
 }
 
-Vec3& Vec3::operator*=(const double d) {
+Vec3& Vec3::operator*=(const float d) {
     vector[0] *= d;
     vector[1] *= d;
     vector[2] *= d;
     return *this;
 }
 
-Vec3& Vec3::operator/=(const double d) {
+Vec3& Vec3::operator/=(const float d) {
     vector[0] /= d;
     vector[1] /= d;
     vector[2] /= d;
     return *this;
 }
 
-double Vec3::length() { return std::sqrt(lenght_sqrd()); }
+float Vec3::length() { return std::sqrt(lenght_sqrd()); }
 
-double Vec3::length_sqrd() { 
+float Vec3::length_sqrd() { 
     return std::pow(vector[0], 2) + 
 	   std::pow(vector[1], 2) + 
 	   std::pow(vector[2], 2);
@@ -92,45 +92,45 @@ Vec3 Vec3::operator/(const Vec3& u, const Vec3& v){
 		      u[2] / v[2]);
 }
 
-Vec3 Vec3::operator+(double d, const Vec3& v){
+Vec3 Vec3::operator+(float d, const Vec3& v){
     return Vec3::Vec3(d + v[0], 
 		      d + v[1], 
 		      d + v[2]);
 }
 
-Vec3 Vec3::operator+(const Vec3& v, double d){
+Vec3 Vec3::operator+(const Vec3& v, float d){
     return d + v;
 }
 
-Vec3 Vec3::operator-(double d, const Vec3& v){
+Vec3 Vec3::operator-(float d, const Vec3& v){
     return Vec3::Vec3(d - v[0], 
 		      d - v[1], 
 		      d - v[2]);
 }
 
-Vec3 Vec3::operator-(const Vec3& v, double d){
+Vec3 Vec3::operator-(const Vec3& v, float d){
     return Vec3::Vec3(v[0] - d, 
 		      v[1] - d, 
 		      v[2] - d);
 }
 
-Vec3 Vec3::operator*(double d, const Vec3& v){
+Vec3 Vec3::operator*(float d, const Vec3& v){
     return Vec3::Vec3(d * v[0], 
 		      d * v[1], 
 		      d * v[2]);
 }
 
-Vec3 Vec3::operator*(const Vec3& v, double d){
+Vec3 Vec3::operator*(const Vec3& v, float d){
     return d * v;
 }
 
-Vec3 Vec3::operator/(const Vec3& v, double d){
+Vec3 Vec3::operator/(const Vec3& v, float d){
     return Vec3::Vec3(v[0] / d, 
 		      v[1] / d, 
 		      v[2] / d);
 }
 
-double Vec3::prod_escalar(const Vec3& u, const Vec3& v){
+float Vec3::prod_escalar(const Vec3& u, const Vec3& v){
     return u[0] * v[0] + 
 	   u[1] * v[1] + 
 	   u[2] * v[2];
